+++ src/test/java/calculator/domain/DelimiterExtractorTest.java
@@ -3,6 +3,7 @@ package calculator.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import calculator.service.DelimiterExtractor;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
